+++ chain/x/oracle/keeper/request.go
@@ -88,8 +88,8 @@ func (k Keeper) SaveResult(ctx sdk.Context, id types.RequestID, status types.Res
 		r.ClientID, id, k.GetReportCount(ctx, id), r.RequestTime,
 		ctx.BlockTime().Unix(), status, result,
 	)
-	resultHash := types.CalculateResultHash(req, res)
-	k.SetResult(ctx, id, resultHash)
+	encodedResult := types.CalculateEncodedResult(req, res)
+	k.SetResult(ctx, id, encodedResult)
 	ctx.EventManager().EmitEvent(sdk.NewEvent(
 		types.EventTypeRequestExecute,
 		sdk.NewAttribute(types.AttributeKeyClientID, req.ClientID),
@@ -103,7 +103,7 @@ func (k Keeper) SaveResult(ctx sdk.Context, id types.RequestID, status types.Res
 		sdk.NewAttribute(types.AttributeKeyRequestTime, fmt.Sprintf("%d", res.RequestTime)),
 		sdk.NewAttribute(types.AttributeKeyResolveTime, fmt.Sprintf("%d", res.ResolveTime)),
 		sdk.NewAttribute(types.AttributeKeyResult, string(res.Result)),
-		sdk.NewAttribute(types.AttributeKeyResultHash, hex.EncodeToString(resultHash)),
+		sdk.NewAttribute(types.AttributeKeyEncodedResult, hex.EncodeToString(encodedResult)),
 	))
 	return res
 }