+++ chain/x/oracle/types/events.go
@@ -30,5 +30,5 @@ const (
 	AttributeKeyResolveTime    = "resolve_time"
 	AttributeKeyResolveStatus  = "resolve_status"
 	AttributeKeyResult         = "result"
-	AttributeKeyResultHash     = "result_hash"
+	AttributeKeyEncodedResult  = "encoded_result"
 )