+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Chain
 
+- (impv) [\#1881](https://github.com/bandprotocol/bandchain/pull/1881) Store result in store instead of result hash.
 - (impv) [\#1901](https://github.com/bandprotocol/bandchain/pull/1901) Added `moniker` field to `delegations_view` table.
 - (feat) [\#1879](https://github.com/bandprotocol/bandchain/pull/1879) Keep `Request` and `Report` around. We avoid premature optimization at the moment.
 - (feat) [\#1873](https://github.com/bandprotocol/bandchain/pull/1873) Add `in-before-resolve` field in `Report` data structure.